Reintegrated graphical data mapping branch
git-svn-id: https://ctf.open.collab.net/svn/repos/ccf/trunk/gui@2933 6771e1e2-0f46-0410-b5d7-d846ff89f743

+++ com.collabnet.ccf/src/com/collabnet/ccf/actions/ChangeSynchronizationStatusAction.java
@@ -30,17 +30,63 @@ public class ChangeSynchronizationStatusAction extends ActionDelegate {
 			if (object instanceof SynchronizationStatus) {
 				SynchronizationStatus status = (SynchronizationStatus)object;
 				File xslFile = status.getXslFile();
+				File graphicalXslFile = status.getGraphicalXslFile();
+				File sourceRepositorySchemaFile = status.getSourceRepositorySchemaFile();
+				File targetRepositorySchemaFile = status.getTargetRepositorySchemaFile();
+				File genericArtifactToSourceRepositorySchemaFile = status.getGenericArtifactToSourceRepositorySchemaFile();
+				File genericArtifactToTargetRepositorySchemaFile = status.getGenericArtifactToTargetRepositorySchemaFile();
+				File sourceRepositorySchemaToGenericArtifactFile = status.getSourceRepositorySchemaToGenericArtifactFile();
+				File targetRepositorySchemaToGenericArtifactFile = status.getTargetRepositorySchemaToGenericArtifactFile();
+				
 				ChangeProjectMappingDialog dialog = new ChangeProjectMappingDialog(Display.getDefault().getActiveShell(), status);
 				if (dialog.open() == ChangeProjectMappingDialog.CANCEL) return;
 				if (!projectMappingsList.contains(status.getProjectMappings())) {
 					projectMappingsList.add(status.getProjectMappings());
 				}
-				if (xslFile != null && xslFile.exists() && dialog.isXslFileNameChanged()) {
+				
+				if (dialog.isXslFileNameChanged()) {
 					File newXslFile = new File(xslFile.getParentFile(), dialog.getNewXslFileName());
-					if (!newXslFile.exists()) {
-						if (MessageDialog.openQuestion(Display.getDefault().getActiveShell(), "Change Project Mapping", "Do you wish to rename field mapping file from " + dialog.getOldXslFileName() + " to " + dialog.getNewXslFileName() + "?")) {
-							xslFile.renameTo(newXslFile);
-						}
+					File newGraphicalXslFile = new File(graphicalXslFile.getParentFile(), dialog.getNewGraphicalXslFileName());
+					File newSourceRepositorySchemaFile = new File(sourceRepositorySchemaFile.getParentFile(), dialog.getNewSourceRepositorySchemaFileName());
+					File newTargetRepositorySchemaFile = new File(targetRepositorySchemaFile.getParentFile(), dialog.getNewTargetRepositorySchemaFileName());
+					File newGenericArtifactToSourceRepositorySchemaFile = new File(genericArtifactToSourceRepositorySchemaFile.getParentFile(), dialog.getNewGenericArtifactToSourceRepositorySchemaFileName());
+					File newGenericArtifactToTargetRepositorySchemaFile = new File(genericArtifactToTargetRepositorySchemaFile.getParentFile(), dialog.getNewGenericArtifactToTargetRepositorySchemaFileName());
+					File newSourceRepositorySchemaToGenericArtifactFile = new File(sourceRepositorySchemaToGenericArtifactFile.getParentFile(), dialog.getNewSourceRepositorySchemaToGenericArtifactFileName());
+					File newTargetRepositorySchemaToGenericArtifactFile = new File(targetRepositorySchemaToGenericArtifactFile.getParentFile(), dialog.getNewTargetRepositorySchemaToGenericArtifactFileName());
+					if ((xslFile != null && xslFile.exists() && !newXslFile.exists()) ||
+						(graphicalXslFile != null && graphicalXslFile.exists() && !newGraphicalXslFile.exists()) ||
+						(sourceRepositorySchemaFile != null && sourceRepositorySchemaFile.exists() && !newSourceRepositorySchemaFile.exists()) ||
+						(targetRepositorySchemaFile != null && targetRepositorySchemaFile.exists() && !newTargetRepositorySchemaFile.exists()) ||
+						(genericArtifactToSourceRepositorySchemaFile != null && genericArtifactToSourceRepositorySchemaFile.exists() && !newGenericArtifactToSourceRepositorySchemaFile.exists()) ||					
+						(genericArtifactToTargetRepositorySchemaFile != null && genericArtifactToTargetRepositorySchemaFile.exists() && !newGenericArtifactToTargetRepositorySchemaFile.exists()) ||					
+						(sourceRepositorySchemaToGenericArtifactFile != null && sourceRepositorySchemaToGenericArtifactFile.exists() && !newSourceRepositorySchemaToGenericArtifactFile.exists()) ||			
+						(targetRepositorySchemaToGenericArtifactFile != null && targetRepositorySchemaToGenericArtifactFile.exists() && !newTargetRepositorySchemaToGenericArtifactFile.exists())) {		
+						if (MessageDialog.openQuestion(Display.getDefault().getActiveShell(), "Change Project Mapping", "Do you wish to rename field mapping files?")) {
+							if (xslFile != null && xslFile.exists() && !newXslFile.exists()) {
+								xslFile.renameTo(newXslFile);
+							}
+							if (graphicalXslFile != null && graphicalXslFile.exists() && !newGraphicalXslFile.exists()) {
+								graphicalXslFile.renameTo(newGraphicalXslFile);
+							}
+							if (sourceRepositorySchemaFile != null && sourceRepositorySchemaFile.exists() && !newSourceRepositorySchemaFile.exists()) {
+								sourceRepositorySchemaFile.renameTo(newSourceRepositorySchemaFile);
+							}
+							if (targetRepositorySchemaFile != null && targetRepositorySchemaFile.exists() && !newTargetRepositorySchemaFile.exists()) {
+								targetRepositorySchemaFile.renameTo(newTargetRepositorySchemaFile);
+							}
+							if (genericArtifactToSourceRepositorySchemaFile != null && genericArtifactToSourceRepositorySchemaFile.exists() && !newGenericArtifactToSourceRepositorySchemaFile.exists()) {
+								genericArtifactToSourceRepositorySchemaFile.renameTo(newGenericArtifactToSourceRepositorySchemaFile);
+							}	
+							if (genericArtifactToTargetRepositorySchemaFile != null && genericArtifactToTargetRepositorySchemaFile.exists() && !newGenericArtifactToTargetRepositorySchemaFile.exists()) {
+								genericArtifactToTargetRepositorySchemaFile.renameTo(newGenericArtifactToTargetRepositorySchemaFile);
+							}	
+							if (sourceRepositorySchemaToGenericArtifactFile != null && sourceRepositorySchemaToGenericArtifactFile.exists() && !newSourceRepositorySchemaToGenericArtifactFile.exists()) {
+								sourceRepositorySchemaToGenericArtifactFile.renameTo(newSourceRepositorySchemaToGenericArtifactFile);
+							}	
+							if (targetRepositorySchemaToGenericArtifactFile != null && targetRepositorySchemaToGenericArtifactFile.exists() && !newTargetRepositorySchemaToGenericArtifactFile.exists()) {
+								targetRepositorySchemaToGenericArtifactFile.renameTo(newTargetRepositorySchemaToGenericArtifactFile);
+							}								
+						}					
 					}
 				}
 				mappingsChanged = true;
